Doing this commit in case somethin goes wrong
Renamed my main scene to "Level 1". The only scene in my scene folder in unity was of a previous save of my project, and I thought for a moment that my progress had been erased. I found and fixed the issue though, so I'm doing this commit as a fail safe in case something happens